pyproject.toml
[project]
name = "mlx-ppocr"
version = "0.1.0"
description = "PP-OCRv5 server on Apple MLX"
requires-python = ">=3.10"
dependencies = [
    "mlx>=0.25.0",
    "numpy>=1.24.0",
    "pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "pyclipper>=1.3.0",
    "shapely>=2.0.0",
    "safetensors>=0.4.0",
    "huggingface-hub>=0.20",
]

[project.scripts]
mlx-ocr = "mlx_ppocr.cli:main"

[project.optional-dependencies]
multilingual = ["paddlepaddle>=3.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.9",
    "torch>=2.0",
    "transformers>=4.50",
    "accelerate>=1.0",
    "torchvision>=0.18",
]
